requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
python-dotenv>=1.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
jinja2>=3.1.0